added new Features

+++ app/Http/Controllers/Topic.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Modules;
 use App\Models\RecentActivity;
 use App\Models\Topics;
 use Illuminate\Http\Request;
@@ -11,28 +12,33 @@ class Topic extends Controller
 
     // ADD TOPIC
 
-    public function addTopic(Request $request){
+    public function addModule(Request $request){
         $input = $request->validate([
-            'subject_id' => 'required|integer',
-            'topic_name' => 'required|string|max:255'
+            'subject_id' => 'required|uuid',
+            'module_name' => 'required|string|max:255'
         ]);
 
-        Topics::create([
+        $maxOrder = Modules::where('subject_id', $input['subject_id'])->max('order');
+        $nextOrder = is_null($maxOrder) ? 1 : $maxOrder + 1;
+        $topics = Modules::create([
+            'order' => $nextOrder,
             'subject_id' => $input['subject_id'],
-            'topic_name' => $input['topic_name']
+            'module_name' => $input['module_name']
         ]);
 
         RecentActivity::create([
             'name' => auth()->user()->name,
-            'action' => 'added_topic',
+            'action' => 'added_module',
             'subject' => $request->subject_name,
-            'topic' => $input['topic_name']
+            'module' => $input['module_name']
         ]);
 
         return response()->json([
             'success' => true,
-            'message' => 'Topic added Successfully'
+            'message' => 'Topic added Successfully',
+            'new_topics' => $topics
         ]);
+
     }
 
     public function editTopic(Request $request){
@@ -59,9 +65,19 @@ class Topic extends Controller
 
     }
 
-    public function deleteTopic(Request $request){
-
+    public function deleteTopic(Request $request)
+    {
         $topic = Topics::find($request->topic_id);
+
+        if (!$topic) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Topic not found!'
+            ], 444);
+        }
+
+        $topicName = $topic->topic_name;
+        $topicId = $topic->id;
 
         $topic->delete();
 
@@ -69,12 +85,13 @@ class Topic extends Controller
             'name' => auth()->user()->name,
             'action' => 'deleted_topic',
             'subject' => $request->subject_name,
-            'topic' => $topic->topic_name
+            'topic' => $topicName
         ]);
 
         return response()->json([
             'success' => true,
-            'message' => 'Topic deleted Successfully!'
+            'message' => 'Topic deleted Successfully!',
+            'deleted_topic_id' => $topicId
         ]);
     }
 }
